requirements.txt
absl-py==1.1.0
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
filelock==3.7.1
fonttools==4.33.3
gast==0.5.3
gdown==4.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
joblib==1.1.0
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
